Revert "Chỉnh Undo Redo"
