--- MyPaint/MyPaint/PaintFormer/ShapeFormer.cs
@@ -1,46 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Paint
-{
-    public class ShapeFormer
-    {
-        //Tạo hình chữ nhật tại các tọa độ chuột
-        public Rectangle FormRectangle(Point ptMouseDown, Point ptCurrent)
-        {
-            return new Rectangle(
-                Math.Min(ptMouseDown.X, ptCurrent.X),
-                Math.Min(ptMouseDown.Y, ptCurrent.Y),
-                Math.Abs(ptMouseDown.X - ptCurrent.X),
-                Math.Abs(ptMouseDown.Y - ptCurrent.Y));
-        }
-
-        //Tạo hình vuông tại các tọa độ chuột
-        public Rectangle FormSquare(Point ptMouseDown, Point ptCurrent)
-        {
-            if ((ptMouseDown.X - ptCurrent.X) > Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)))
-                return new Rectangle(
-                    ptMouseDown.X - Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)),
-                        Math.Min(ptMouseDown.Y, ptCurrent.Y),
-                        Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)),
-                        Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)));
-
-            else if ((ptMouseDown.Y - ptCurrent.Y) > Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)))
-                return new Rectangle(
-                    Math.Min(ptMouseDown.X, ptCurrent.X),
-                    ptMouseDown.Y - Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)),
-                    Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)),
-                    Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)));
-
-            return new Rectangle(
-                Math.Min(ptMouseDown.X, ptCurrent.X),
-                Math.Min(ptMouseDown.Y, ptCurrent.Y),
-                Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)),
-                Math.Min(Math.Abs(ptCurrent.X - ptMouseDown.X), Math.Abs(ptMouseDown.Y - ptCurrent.Y)));
-        }
-    }
-}
